Moving credentials from internal to pkg
