+++ internal/commands/login.go
@@ -26,10 +26,10 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/docker/hub-tool/internal/ansi"
-	"github.com/docker/hub-tool/internal/credentials"
 	"github.com/docker/hub-tool/internal/errdef"
 	"github.com/docker/hub-tool/internal/login"
 	"github.com/docker/hub-tool/internal/metrics"
+	"github.com/docker/hub-tool/pkg/credentials"
 	"github.com/docker/hub-tool/pkg/hub"
 )
